# v1.1/backend/docs/swagger.yaml
definitions:
  analyze.SpaceCategoryAnalyzeRequest:
    properties:
      queryAll:
        description: 是否查询所有空间
        type: boolean
      queryPublic:
        description: 是否查询公开空间
        type: boolean
      spaceId:
        description: 空间ID
        example: ""
        type: string
    type: object
  analyze.SpaceCategoryAnalyzeResponse:
    properties:
      category:
        description: 分类名称
        type: string
      count:
        description: 分类数量
        type: integer
      totalSize:
        description: 分类总大小
        type: integer
    type: object
  analyze.SpaceRankAnalyzeRequest:
    properties:
      top_n:
        description: 排名前N的空间
        type: integer
    type: object
  analyze.SpaceTagAnalyzeRequest:
    properties:
      queryAll:
        description: 是否查询所有空间
        type: boolean
      queryPublic:
        description: 是否查询公开空间
        type: boolean
      spaceId:
        description: 空间ID
        example: ""
        type: string
    type: object
  analyze.SpaceTagAnalyzeResponse:
    properties:
      count:
        description: 标签数量
        type: integer
      tag:
        description: 标签名称
        type: string
    type: object
  analyze.SpaceUsageAnalyzeRequest:
    properties:
      queryAll:
        description: 是否查询所有空间
        type: boolean
      queryPublic:
        description: 是否查询公开空间
        type: boolean
      spaceId:
        description: 空间ID
        example: ""
        type: string
    type: object
  analyze.SpaceUsageAnalyzeResponse:
    properties:
      countUsageRatio:
        description: 资源数量使用比例
        type: number
      maxCount:
        description: 最大资源数量
        type: integer
      maxSize:
        description: 最大空间大小
        type: integer
      sizeUsageRatio:
        description: 空间使用比例
        type: number
      usedCount:
        description: 已使用的资源数量
        type: integer
      usedSize:
        description: 已使用的空间大小
        type: integer
    type: object
  analyze.SpaceUserAnalyzeRequest:
    properties:
      queryAll:
        description: 是否查询所有空间
        type: boolean
      queryPublic:
        description: 是否查询公开空间
        type: boolean
      spaceId:
        description: 空间ID
        example: ""
        type: string
      timeDimension:
        description: 时间维度：day/week/month
        type: string
      userId:
        description: 用户ID
        example: ""
        type: string
    type: object
  analyze.SpaceUserAnalyzeResponse:
    properties:
      count:
        description: 周期内上传的图片数量
        type: integer
      period:
        description: 时间周期
        type: string
    type: object
...
  common.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
...
  entity.Space:
    properties:
      createTime:
        type: string
      editTime:
        type: string
      id:
        example: ""
        type: string
      maxCount:
        type: integer
      maxSize:
        type: integer
      spaceLevel:
        type: integer
      spaceName:
        type: string
      spaceType:
        type: integer
      totalCount:
        type: integer
      totalSize:
        type: integer
      updateTime:
        type: string
      userId:
        example: ""
        type: string
    type: object
  entity.SpaceUser:
    properties:
      createTime:
        type: string
      id:
        example: ""
        type: string
      spaceId:
        example: ""
        type: string
      spaceRole:
        type: string
      updateTime:
        type: string
      userId:
        example: ""
        type: string
    type: object
...
  space.SpaceVO:
    properties:
      createTime:
        type: string
      editTime:
        type: string
      id:
        description: Space ID
        example: ""
        type: string
      maxCount:
        type: integer
      maxSize:
        type: integer
      permissionList:
        description: 空间的权限列表
        items:
          type: string
        type: array
      spaceLevel:
        type: integer
      spaceName:
        type: string
      spaceType:
        description: 'Space type: 0 - 私人空间, 1 - 团队空间'
        type: integer
      totalCount:
        type: integer
      totalSize:
        type: integer
      updateTime:
        type: string
      user:
        $ref: '#/definitions/user.UserVO'
      userId:
        description: User ID
        example: ""
        type: string
    type: object
  spaceuser.SpaceUserAddRequest:
    properties:
      spaceId:
        description: 空间ID
        example: ""
        type: string
      spaceRole:
        description: 空间角色：viewer-查看者 editor-编辑者 admin-管理员
        type: string
      userId:
        description: 用户ID
        example: ""
        type: string
    type: object
  spaceuser.SpaceUserEditRequest:
    properties:
      Id:
        description: 表的元组ID
        example: ""
        type: string
      spaceRole:
        description: 空间角色：viewer-查看者 editor-编辑者 admin-管理员
        type: string
    type: object
  spaceuser.SpaceUserQueryRequest:
    properties:
      Id:
        description: 表的元组ID
        example: ""
        type: string
      spaceId:
        description: 空间ID
        example: ""
        type: string
      spaceRole:
        description: 空间角色：viewer-查看者 editor-编辑者 admin-管理员
        type: string
      userId:
        description: 用户ID
        example: ""
        type: string
    type: object
  spaceuser.SpaceUserRemoveRequest:
    properties:
      Id:
        description: 表的元组ID
        example: ""
        type: string
    required:
    - Id
    type: object
  spaceuser.SpaceUserVO:
    properties:
      createTime:
        type: string
      id:
        example: ""
        type: string
      space:
        allOf:
        - $ref: '#/definitions/space.SpaceVO'
        description: 空间信息
      spaceId:
        example: ""
        type: string
      spaceRole:
        type: string
      updateTime:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/user.UserVO'
        description: 用户信息
      userId:
        example: ""
        type: string
    type: object
...
  user.UserVO:
    properties:
      createTime:
        type: string
      id:
        example: ""
        type: string
      userAccount:
        type: string
      userAvatar:
        type: string
      userName:
        type: string
      userProfile:
        type: string
      userRole:
        type: string
    type: object
host: localhost:8001
info:
  contact:
    email: lwh24621@qq.com
    name: lwhhhh
    url: 无
  description: CanvasCloud
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: 无服务条款
  title: CanvasCloud
  version: "1.0"
paths:
...
  /v1/space/analyze/category:
    post:
      consumes:
      - application/json
      parameters:
      - description: 查询条件
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/analyze.SpaceCategoryAnalyzeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/analyze.SpaceCategoryAnalyzeResponse'
                  type: array
              type: object
        "400":
          description: 更新失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 获取空间图片分类分析「登录校验」
      tags:
      - space/analyze
  /v1/space/analyze/rank:
    post:
      consumes:
      - application/json
      parameters:
      - description: 查询条件
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/analyze.SpaceRankAnalyzeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.Space'
                  type: array
              type: object
        "400":
          description: 更新失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 获取空间使用情况排名「管理员」
      tags:
      - space/analyze
  /v1/space/analyze/tag:
    post:
      consumes:
      - application/json
      parameters:
      - description: 查询条件
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/analyze.SpaceTagAnalyzeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/analyze.SpaceTagAnalyzeResponse'
                  type: array
              type: object
        "400":
          description: 更新失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 获取空间标签出现量分析「登录校验」
      tags:
      - space/analyze
  /v1/space/analyze/usage:
    post:
      consumes:
      - application/json
      parameters:
      - description: 查询条件
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/analyze.SpaceUsageAnalyzeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/analyze.SpaceUsageAnalyzeResponse'
              type: object
        "400":
          description: 更新失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 获取空间使用分析「登录校验」
      tags:
      - space/analyze
  /v1/space/analyze/user:
    post:
      consumes:
      - application/json
      parameters:
      - description: 查询条件
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/analyze.SpaceUserAnalyzeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/analyze.SpaceUserAnalyzeResponse'
                  type: array
              type: object
        "400":
          description: 更新失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 获取用户上传图片统计分析，支持分析特定用户「登录校验」
      tags:
      - space/analyze
...
  /v1/spaceUser/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: 成员的ID和空间ID，以及添加的成员角色
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/spaceuser.SpaceUserAddRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 返回空间成员表的数据ID，字符串格式
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 查询失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 增加成员到空间
      tags:
      - spaceUser
  /v1/spaceUser/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 空间成员表中数据的ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/spaceuser.SpaceUserRemoveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功与否
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: 查询失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 从空间移除成员
      tags:
      - spaceUser
  /v1/spaceUser/edit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 记录的ID和需要调整的权限
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/spaceuser.SpaceUserEditRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 编辑成功与否
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: 查询失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 编辑成员权限
      tags:
      - spaceUser
  /v1/spaceUser/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: 必须携带spaceID和userID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/spaceuser.SpaceUserQueryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功与否
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.SpaceUser'
              type: object
        "400":
          description: 查询失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 查询某个成员在某个空间的信息
      tags:
      - spaceUser
  /v1/spaceUser/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 可以携带的参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/spaceuser.SpaceUserQueryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 返回详细数据
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/spaceuser.SpaceUserVO'
                  type: array
              type: object
        "400":
          description: 查询失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 查询成员信息列表
      tags:
      - spaceUser
  /v1/spaceUser/list/my:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 返回详细数据
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/spaceuser.SpaceUserVO'
                  type: array
              type: object
        "400":
          description: 查询失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 查询我加入的团队空间列表
      tags:
      - spaceUser
...
securityDefinitions:
  BearerAuth:
    description: 'JWT 认证格式: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"